style: do not wrap text in sidebar
This has been discovered as "mysterious" difference between production
deployment and staging/development. It is still an open question why
this happen. There seem to be no difference in any loaded resource by
the web-browser (when checking out the deployed version).

Anyhow, the production behavior is actually the more desired one. At
least based on the current sidebar design. Therefore it got explicitly
styled this way now. This nicely solves the issue of end-to-end tests
running again production (which reported a screenshot diff).

Notice that for unspecific reasons, some screenshots shrunk in size
(4px). Therefore the baseline received a bigger update in this commit.