refactor(smoke-test): fetch deployed commit from test target
Implement a Playwright script to properly load the page and
read the commit hash from the window variable. Initially, I
attempted to solve this with a bash script, but it became too
complex when trying to use it for both staging and production.

This action isn’t a typical JavaScript action with bundled code
because a) It’s not necessary for a local action, and b) We need
to set up Playwright anyway, so by using a yaml action, we can
reuse the cached-playwright-setup action.