fix: conceptual issue on calculation of Elterngeldbezüge
This fixes a very long living bug. At least till the current team
received the code (without history). And from how it looks like, it
exists like this since the conception phase of the old calculator with
planner.

The old application used to calculate all Lebensmonate in advance, based
on the Einkommen during this period of time. And this "simulation" (how
it was called) was continued to be used, based on the assumption that
this just works correctly.
But the calculation of the Elterngeldbezüge depends on the actual Plan.
Einkommen is separated by Elterngeldart/-variante and leads to
different values for the months. It matters if an user has an Einkommen
and chose either Basiselterngeld, EltengeldPlus or Partnerschaftsbonus.
The resulting Elterngeldbezug might vary depending on all the other
planned months with their Optionen and Einkommen.

Luckily this actually puts much less requirements on the calculation
feature. The whole "simulation" part was just on top and can be safely
removed soon. It still needs a test harness that allows more secure
refactorings.

The solution requires to have planning and calculation to be much more
entangled. This gets reflected by shifting the responsibility into the
domain of the planer. In the past, it was the job of the UI services to
update the Elterngeldbezüge after certain user interactions. This is no
more sufficient and is no service logic. That becomes especially obvious
when looking at how Auswahlmöglichkeiten are defined.
Anyhow, it is still highly valuable to separate the calculation
implementation from the planer. They have separate boundaries with their
own domain, having a singly different angle of the topic of Elterngeld.
There requirement of the planer domain to calculate Elterngeldbezüge
gets explicitly requested by simple, plain function arguments. Their
order was exclusively chosen for currying purposes.

This big-ish commit could have potentially be broken down into smaller
changes. But it would have required to establish mechanisms that
maintain the old behavior, at least partially. The trade off was not
worth it and it was rather more simple to just do it all at once. In the
end it was still a quite limited concept about how the calculation is
bound into the planer.