ci: update runner image version
The GitHub Actions CI produced some warnings that the `-latest` tag for
Ubuntu will be updated to `24.04` soon. So pipelines might break if they
have some OS dependencies that changed between the LTS versions of
Ubuntu.
Anyhow, it is actually a better practice to use explicit versions. In
best case combined with some proper dependency update tool that is also
capable to "see" these versions too.