chore(lint): add linting for GitHub actions workflows
It is too easy to introduce an error/bug into the workflows. In best
case this will be recognize when the next run fails. In any case this is
a long feedback loop, potentially blocking the repository. Therefore
such "configurations" should be treated like regular code, including
linting.

The applied fixes were exclusively for the "sub-linting" of the used
shell commands by shellcheck.

Notice that this puts new requirements on the developer environment to
have the `actionlint` binary available. This is another push towards
a developer environment as code. And it will become more and more
urgent.