chore: remove dead code from Einkommen store
This code was used to calculate the total Einkommen per Elternteil and
identify it that's higher than the maximum.
In the short history of the code we have, we just see that in the last
update we got from the former team, the feature got removed.
Unfortunately all the related code was not removed as well. Which is
actually a lot.

We actually have a ticket related to this topic already open and will
revisit this anyway.