feat(calculation): add custom Programmablaufplan implementations
The calculation of Elterngeld depends on some tax levies calculation. In
the past this was done via two different interfaces. First of all there
was a remote service. Second a library that was bundled with the
application. Both originated from the Bundesfinanzministerium (BMF) who
is in charge of this. Anyhow, the remote service has been discontinued
in our application a while ago (though, its stubs are still there). And
the library does not see any updates anymore. Additionally we have been
informed that it is incorrect anyhow. Because of that, we already can't
calculate proper Elterngeld for many years of birth, especially those
for 2025 (but also older ones).
The BMF announced to start an initiative, whereof of well designed and
maintained library/service will emerge from again. But this will take
some time. In the meantime we are on our own. Actually we are already
for a long time for all the parents planning to give birth 2025 or
parents who reapply late.

Most important informations are as comments in the code actually. This
is a special "design" and requires in-code documentation and
explanation.
So far, this is just the "kick-off". It is not yet used by the
application. That will come in steps soon to come.

Happily we can say that all of the 100% fundamental test cases of the
BMF are successful. The currently still used library only hits about 30%
of the cases, whereof ~10% are trivial cases.