chore: remove unknown Steuerklasse
This does not make sense within the domain of the calculation. There is
no unknown Steuerklasse in this context. Any uncertain state must be
sorted out before entering this domain.
That is currently the case for the query section with the user. To keep
this change terse, the Steuerklasse 1 is used as sane default during
parsing now.