feat(calculcation): inject custom tax calculation
Notice that this causes (again) huge issues with our test suite. In all
test cases, the differences are within a few cent or Euro. Rarely much
more. From the user perspective this should be fine for now.
Anyhow, our test suite that test again the spreadsheet of the FIT
institute cause more trouble. We need to figure out how to maintain this
in future. Especially in focus of multiple tax and law years.

This does not yet do the clean up of all the obsolete stuff. Which is
actually quite a lot. Especially because there is also the remote
service implementation too. It should no shadow the changes in this
commit.