fix(calculation): control side-effects of input data
The reported bug was, that when having Mischeinkommen, the Planer shows
only the minimum amount of Elterngeldbezug.
Turned out the underlying issue is caused by the calculation not being
side-effect free. More specifically it manipulates its input data in
a way when it is called again with the manipulated data, it results into
the minimum values only.
This happened, because after the question section, the data for the
calculation is mostly cached. Except for the data of the planer. In
result, after the very first calculation this issue happens. And because
how options are "simulated", this happens already for the second option
of the first Lebensmonat.

The underlying issue should be properly addressed. And in future
prevented by using immutable input data structures. But that isn't
easily possible right now. Especially because the tests don't provide
the security to detect side-effects on refactoring.