feat(monatsplaner): establish strong typing for pseudonyms
This commit has turned into a much bigger refatoring than originally
anticipated. But its for a good cause.

The origin of this commit is the strengthening of the Ausgangslage type.
This type is in the seed of the Monatsplaner. It theoretically dictates
the Monatsplaner its central polymorphism parameter: the number of
Elternteile. Theoretically, because it wasn't properly implemented this
way before the commit.
Despite the strong generic typing across the whole Monatsplaner, there
are partially different behaviors depending on the number of
Elternteile. So it is the case for the Pseudonyme. Though, it wasn't
properly described in the type system before.

Pseudonyme for Elternteile are used to distinguish between them in the
user interface. It is used for table headers, many aria labels and
descriptions, and more. Though, the don't act as identifiers/keys.
In case of a single Elternteil, there is no need for this. And the whole
user interface adapts slightly to this case.
Before this commit, the Pseudoynm has been set to an empty string. Which
is not very helpful and expressive from the perspective of type system.
A rather simple approach would have been to just set the type of
a string to null/undefined in case of a single Elternteil. But actually
there are no Pseudonyme at all in this case. The latter is even more
"screaming" when using, making this fact more obvious. So the developer
is more strongly forced to handle this case properly.

Unfortunately is has turned out that the Pseudonyme have been somewhat
abused. The Plan, or rather the Lebensmonate, don't provide the
possibility to iterate over the Elternteile. But this is often needed to
display UI elements per Elternteil. So the Pseudonyme where a reliable
source. This is no more possible with the new type definition.
Therefore, this information is derived from the Ausgangslage now. As
this is the source of truth, from a type and data perspective.
In result it was necessary to drill down the whole Ausgangslage. Before
this, parts of the Ausgangslage, like the Pseudonyme, have been
forwarded only. This has shown how important this information (which
Elternteile are there) actually become, but was covered by suboptimal
code.
Many more parts in the UI have been adapted, where rendering per
Elternteil was based on weak data structures which don't guarantee
to be reliable for this purpose. In fact it is actually planned to
loosen them even more.

It is not quite optimal to pass down the whole Ausgangslage too deep.
Before this, components could express their dependencies more detailed
(e.g. only need the Pseudonyme). The Ausgangslage holds quite a sum of
data. Whereof plenty is not of interest for any user interface
component. Unfortunately, as the type is defined right now, it can't be
easily separated without loosing the strong type guarantees. This is
first and for most the relation between the number of Elternteile and
the Pseudonyme. But also other properties depend on the number of
Elternteile.
For the moment, this is acceptable. As the type is designed right now,
the majority of properties is optional. This lowers the burden on tests
to create sample data (as we still have no factories). There is also no
secret/hidden data, even cross feature boundaries. Anyhow, we might come
up with an alternative approach to improve and use more reduced data.
Ideas are already iterated.