chore(monatsplaner): slightly improve import usage
Expressing architecture "rules" via visibility modifiers cross modules
is not possible in TypeScript. This needs other mechanisms.

The planer feature is split into the domain and the user interface. The
latter is separated into the service that provides the domain and the
components that consume the service. Finally the top level components,
the Planer and Zusammenfassung are meant to be used as single entry
points to the feature.

For simplicity, the service layer does transparently use the types of
the domain. Furthermore, the domain provides a bunch of neat utility
functions to "read" the data in a type safe manner. In result, plenty of
components from the user interface depend on the domain too. And on the
next level, users of the whole feature also started to depend on them.

This is not necessarily too bad right now. Though, it would be probably
nice to break down in future. At least at the feature boundary. But this
needs a lot of work to break down the code base into compiler bundles to
enforce this properly. This isn't worth the effort yet. At least as long
as developers work carefully, with the intended design in mind.

Anyhow, there was once the pattern established that the user interface
components should only depend on the service. Which made the service
re-export tons of stuff from the domain. Which was just annoying. And it
was too easy to "break" this rule by automatic IDE imports.
Therefore, this has been resolved and direct dependencies are used. This
makes the flaw very visible and does not "hide" it. But it also leaves
the code base a little more clean and less confused. Proper architecture
enforcements might follow in future.