ci: improve how the refactoring property tests of the calculation are run
This adds two conditions to the tests. First of all, it runs the test
only if extremely slow tests are allowed. Because there is no better
mechanism, a plain simple environment variable is used. Unfortunately
does the `only` option of Vitest not help for this scenario.

Furthermore, the number of property test runs varies depending on
running in a continuous integration pipeline or not. So it reduces the
number of runs to achieve a reasonable execution duration for the
pipeline. For local execution, the number of runs higher to improve the
reliability of the tests.