test: improve developer experience for refactoring property tests
It seems to become a complex task to transform the output data of the
Elterngeld calculations to comparison. Due to refactoring, the format
changes slightly and transformations need to be done carefully. To allow
this in a more controlled and "secure" manner, a new concept has been
implemented. It should help to control the complexity and to avoid
mistakes.