refactor(calculation): remove obsolete calculation methods
Due to a recent fix of how Elterngeld gets calculated for the Planer,
some calculation methods have become obsolete.
One test case could be split for the two Elternteile, as Elterngeld can
be only calculated per Elternteil individually now. Another test case
must be removed as it is basically a manifestation of the formerly wrong
behavior.