feat(calculation): add Programmablaufplan for taxes in 2024
This is the first time the overarching interface has been adapted. The
documentation was slightly improved by the learnings.

Notice that the usage of the PVA input parameter is yet unclear. It
needs to be resolved with the project partners and domain experts if
Geschwister as of the BEEG law can be applied here.