style: replace px-to-rem function with utility spacing
This commit is in the context of establishing uniform styling.

Some spacing had to be changed to align with the style scheme. But the
majority could be mapped 1:1.

It has been shown that plenty of styles/classes/... are undetected
obsolete. Only a tiny minority has been resolved. The commit tries to stay
focused on this one task to remove the `px-to-rem` usage.