style: inline media breakpoint variables
These breakpoints are no good and solid design. We also got
accessibility issues reported for them. The Sidebar component (which
"bleeds" into the Page component) and the InfoDialog are the last
components struggling with this issue.
To get rid of the undesired globals SCSS variables of breakpoints, they
got inlined for now. This does not really make it better or worse. We
definitely wanna resolve these breakpoint soonish.