style: resolve namespace injection
This is in the context of unifying the styling.

The SCSS part of styling uses BEM class names. All class names are
scoped in a namespace. In general, this is a good approach. Especially
when being not a standalone application.
Anyhow, the approach how each class name gets defined in SCSS and each
class usage in TypeScript is quite over engineered. Everything to allow
replacing the namespace.
First of all, this is not really a use-case. And in the unlikely case it
should actually happen, the costs to resolve this are fairly low.
Especially with the security of visual regression tests in place.
Second, this makes it even harder to follow the styling. This styling
approach decouples already the document structure from the styling in
separate files. Furthermore, variables are somehow exported into another
file. But the worst part is, that the class names in the document part
are hard to match with the stylesheets, as the namespace "code" is in
the way (plus more).
Moreover, the way it is implemented is quite tedious and cumbersome.
A lot of noise in the code.
After all, we intend to refactor all SCSS code anyway. So this change is
for good. It also cleans up the final bit of global SCSS import-stuff.
The same goes similarly for the TypeScript part.