style: simplify SCSS file import structure
The goal is to unify the styling approaches in the code base. In this
context, the SCSS usage for individual components should be removed.
After all `*.variables.scss` files could have been resolved, the purpose
of all the `index.scss` plus `<component_name>.scss` files for each
component have become quite questionable. To reduce the noise and keep
things focused, the index files got removed.

A note on using `index` files. In SCSS as well as TypeScript/JavaScript,
they allow to shorten import paths. But much more importantly, they
allow to hide internal structures. This provides independence and
flexibility.
All of this is not necessary or desirable for the components SCSS
stylesheets. Therefore, to keep the Git history easier to follow, the
`<component_name>.scss` files have been kept, instead of renaming them
all to `index.scss`.

Furthermor, notice the small change to the files in the `applications`
directory. The "root" `index.scss` has been moved to its peers of the
CSS styles into the folder with obvious name. This left only the
`App.tsx` component in the `applications`. The plural of this name makes
no sense for us. Also keeping this directory for a single file does not
express any design pattern or make the architecture more obvious.
Therefore, it has been moved to the root source directory, as it is
quite common.