test: move setup script to utilities directory
Having tests close to their code they test is a good approach. That's
why we value in-source tests so much. Anyhow, managing test utilities
within the root source directory is a different thing. It doesn't feel
right. Anyhow, there is no obvious place to put them else right now. But
we already have a dedicated folder for all this stuff. So it has a nice
label and everyone looking at the directory structure knows what to
expect. The setup script for the tests was just missing to be moved
there. Yes, it is somewhat an init/root/... script for the tests. But
not really. So there is not much value in leaving it there in terms of
screaming architecture. In result, the root source directory is pretty
clean of individual files and easy to grasp. There are tons of
directories (which is another story) to understand when getting familiar
with the code base.