chore: move Elternteil type to state management
The `globals/js` was/is a "lost+found" directory where things just get
parked without any structural meaning. Anyhow, the application is in
a weird state, structure wise. But the Elternteil type here is strongly
related to the "Abfrageteil" feature. It is for example not the same as
the one used for the Planer feature. It is also not part of the Rechner
(`calculations`) feature. But the "Abfrageteil" has no proper structure
and is somehow distributed all other the place. Though, the state
management (the Redux slices) basically define the data models that
this part of the application acts on. Unfortunately, it uses the
Elternteil tye completely implicitly. That's why this movement does not
seem obvious. However, all components and data factories are dependent
on this hierarchy. So this file got moved there and is no more lost.