refactor: move parsing date strings to source
The handling of Dates in the Abfrageteil is horrible using strings,
being empty strings per default, etc pp. Anyhow, there was still
a function to parse such date strings lingering around in the global
utilities directory. Under the effort to resolve this directory, the
function has been co-located to the source that "causes" its usage.
After all, it was barely necessary to export it at all.

Notice that this moves the tracking of the Nutzergruppe to a more
meaningful place. Instead of being dependent on the state and
re-rendering of the related form component, it gets tracked on
successfully submitting the form. Which is much more meaningful and
avoid the date parsing issues to some degree.