feat: properly co-locate formatting of monetary values
This commit is in the context to clean-up the projects directory
structure. One goal is to remove the `util` folder as an anti pattern.

After past refactoring, the `formatAsCurrency` function was solely used
by the Planer feature. Therefore, it belongs there for now. Furthermore,
it serves purely (with one exception) visual purposes. So it was
"converted" into a React component that gets used.
This component potentially moves out of the feature again. But that time
it would likely move to the shared base components.