refactor: properly co-locate the progress step definition
This commit is in the context of cleaning up the projects directory
structure One goal is to remove the `util` folder as it is an anti
pattern.

The from/progress steps are an awkward thing in themselves. They are
indirectly the routes of the application, but also used as progress
indicators and much more. This needs to be resolved. But this has
a broader context that needs proper and big refactoring.
In any case, this has moved to the page components as it is closely
related to them. This becomes more obvious as many relative one-off
imports are used. The usage in the Sidebar and Page components related
to the already mentioned flaws.