refactor: properly co-locate hook for navigation with plan
This commit is in the context of cleaning up the project structure. This
includes to locate things by context and purpose, not by type or name.

The moved hook is shared logic to moved from and between the pages.
Thereby it has also been moved to the pages where it belongs to.