feat: improve usability by better input structure
This makes properly use of fieldsets with legends and sections labeled
by headers.
The latter is a design flaw we wanna resolve soonish. Therefore, it
doesn't matter that there is too much duplicated code after resolving
the FormFieldGroup component.

Some test need to be adapted to the new structure. In first place
because the tests were no good (using test identifier and similar). But
also because there was a hidden conditional aria label used in the past.
The current state of using heading or not based on the number of parents
is no good. We need to resolve this. So right now this is just
a workaround.

While being in first place just structural changes, the baseline of
screenshots for the tests was updated. The diff is absolutely minimal
(after fixing everything else) and actually for the better.

Based on external accessibility report.