fix: disable keyboard navigation for future progress steps
The progress steps allow only to navigate back to the already completed
steps. Unfortunately, it was possible to navigate to any step using the
keyboard.
Disabling links is not properly possible. The links are there and could
be used in some way. To properly resolve this for all possible cases,
only steps that can be navigated to are actually links, everything else
is just pure text. The progress step roles remain the same.

Based on external accessibility report.