fix: dialog accessibility for income limit modal
When the parents are over the income limit, a warning pops up when
opening the planer. This is intended to be a pop modal. But it was just
a fixed element on a high z-index. This isn't a proper modal with
accessibility and usability in mind.
To solve this, a native dialog is used that gets opened as modal. It
automatically solves "all problems" by native browser support.

This warning is planned to removed or rather communicated in a different
way to the user. Likely even somewhere else in the user flow.
Furthermore, the past `ModalPopup` component was used at this single
location only. It is rather annoying to link everything with props,
forward reference etc. pp. In the end, it was much easier to inline
this where used. Having in mind it will become removed soonish.

Notice that based on the usage of a proper modal dialog, the positioning
has changed slightly, as this can't be manipulated apparently.