fix(planer): focus after reset the plan
Theoretically, a minimal fix would have only required to make the
`<div>` element focusable via the `tabIndex` attribute. Anyhow, this
element is no really helpful when using a screen reader. Therefore, the
list of Lebensmonat (a section with a heading) is used instead. This
required reference forwarding, causing the size of the commit.

Based on an external accessibility report.