feat: improve accessibility of sidebar toggle button
The original driver of this change was to add a ARIA label to the toggle
button for better screen reader communication. It should express that
this includes a step progress indication.

The button had declared also ARIA attributes for the expanded state.
Unfortunately, this was not working. The Sidebar used the custom
`Button` component which did not support these attributes as properties.
This was not reported by any linting.
Anyhow, in fact does the whole styling of the Sidebar work against the
Button component. So instead of extending the Button component, just
a simple native button element was used instead.