fix: error messages for Zeitraum inputs
The attempt to mix the errors for the `von` and `bis` inputs of
a Zeitraum does not properly work from an accessibility point of view.
First of all, the ARIA description identifiers where broken. Instead,
the errors get shown as everywhere else for each input field
individually.

Moreover, when the two fields do not make up a valid Zeitraum, the error
should be on the `von` field, as this makes more sense after the form
got validated to jump to this field.

Furthermore, the requirement for both inputs has been drastically
simplified. No idea why it was that complicated before.