fix: proper validation of Erwerbstaetigkeit fieldset
The checkbox inputs of this fieldset have the constraint validated that
at least one of them must be selected. In result, the error does not
belong to a single input. Instead, all inputs are invalid. Plus the
error message is linked to the fieldset. That works because the inputs
belong to the fieldset and get announced together anyway.
It is a little "tricky" where to put the validation. In contrast to how
it was before, it is validated on the first checkbox now. That has the
advantage that after the first validation on submitting the form, the
user automatically focuses the first input of the fieldset and gets it
announced with the fieldset name and description, including the error
message.

Based on an external accessibility report.