refactor(calculations): remove "future" children from persönliche daten
The property `kinder` used to be a mix of the future/expected children
and all the siblings that are already born. The first group of children
always had no disability and no birthdate. Moreover, all "future"
children were always filtered out on any operation acting on the data.
Because the whole algorithm only cares about the siblings.
In result, the data structure was renamed and slightly adapted to
represent this better.