refactor(calculations): convert persönliche Daten class to plain type
The goal is to simplify the interface of the calculation. The current
class can be instantiated with only one of several properties, which
then can be overwritten later on. This is not a great approach. The
defaults used in the constructor are not really necessary. Not having
them, provides more transparent usage. In terms of usage convenience
there are possibilities to further improve this type in future.

In functional style, the data is marked as readonly. But unfortunately
that is not completely possible. This is due to the flaws in the
calculation we are already aware of. It makes this fully visible and
more obvious to recognize now.

Notice that also the cloning operation could have been simplified, as it
is a plain flat object now.