refactor(calculations): remove usage of yes-no enumeration
The goal is to simplify the interface of the calculation. This
enumeration represented true states: yes and no. Every usage within the
calculations feature was for a toggle or feature. Therefore, a primitive
boolean type is much more simple and easier to use. Code gets more
straight forward, easier to read and write, more terse, ... A boolean is
the idiomatic approach to represent this kind of data.

Additionally, all properties/variables which used this type have been
partially renamed to be more idiomatic for flags. Though, only a prefix
has been added, nothing else. Naming is a huge topic on its own in this
part of the application.

Moreover, such flag properties are partially defined as optional now.
The absence of the property represents the negative state. This is
a developer convenience thing and can produce more terse code (e.g. in
the tests).

The whole UI part hasn't been touched. Or rather, the enumeration has
moved into the Redux store directory, where is heavily used to reflect
user inputs for yes and no questions. Here it might make sense to keep
it. Anyhow, for the moment, this keeps the change focused and
"smallish".