feat: improve page layout and navigation
This change was primarily driven by the suboptimal usability, especially
from the accessibility perspective. Keyboard control and screen reader
usage were not well supported when navigating between the pages.

First of all, the former ARIA live region has been dissolved. It was
used to announce the page name on navigation to screen reader. But it
was reported by an accessibility audit to be no good practice. The
suggestion was to rather use existing visually available content to
unify the experience in a robust manner.
Furthermore, the focus was always lost when navigating from one page to
another. In result the screen reader was "lost" and keyboard users had
to tab through the whole navigation etc. every time.
To solve both issues together, when navigating between the pages, the
focus is moved to the top level section element. Thereby, the screen
reader experience is much more fluid. Also keyboard control is much
pleasant.

To achieve the above, the `Page` component had to take over more
responsibility. It was already aware of the current route and navigation
step. Now it is also responsible to render the main section element and
label it with the pages its heading. The page components themselves
are no only responsible for the actual content (below the heading).
The diff of the commit might light partially big due to changed
indentation in some components which got adapted to that change. Also,
because the structure wasn't fully clean beforehand, some headings had
to shift levels due to this. Notice that the Planer is still covered
with sections and headings only available to assistive technology.

Due to the above change, the `text` property of the navigation/progress
steps has been renamed to `heading` as it describe the purpose more
properly. By removing the headings from the page components themselves,
this partially improved the code by having the heading defined in one
location.

Changes in the screenshot test baseline was caused only by a fixed
empty space on the page with the Elterngelddetails.