fix(calculation): add tests to fix recursive data transformation
The refactoring properties test suite relies on some recursive data
transformation to close the gap between both tested algorithms. This
function was not tested and had a bug. Lucky enough, this did not lead
to undetected refactoring issues in the past. Though, it was discovered
during work on the calculations and is necessary for future stability.