fix: wrong calculated bonus deadline by siblings
There were a variety of scenarios with a wrong deadline for the bonus,
calculated based on the siblings birthdates. While the actual logic by
the law is quite simple, the original implementation was unnecessarily
complex.

The basic approach remains the same. Based on the list of siblings, the
youngest, second youngest or youngest disabled sibling is filtered out.
Then, based on the related birthdate, the deadline can be calculated:
the day before the siblings turns 3, 6 or 14 years old. The bonus is
then payed for all Lebensmonate which are before this day or which
includes the day. Nothing more, nothing less.
This has simply be implemented, leading to a hopefully more straight
forward understanding of the code. Also bugs should be much more rare
now due to this.

Furthermore, explicit optional/empty/special-case type has been used.
The deadline was originally already allowed to have a value of `null`
instead of a `Date`. But this was never used. Instead, if there is no
bonus available, the expected birthdate was used. This uses a special
encoded value in the spectrum of valid values. Which is horrible. But
worse: based on the laws logic, when the sibling is born one day after
the kid in the same month and 3, 6 or 14 years before, the deadline is
exactly the kids birthday and will result in a bonus payed out for the
first Lebensmonat.

Moreover, are all these calculations purely on dates without any time.
Unfortunately, this is not natively supported by the standard `Date`
type. It is extremely hard to ensure this, as no type-drive-development
is possible. One approach is to exercise this through at any location in
the code base that works with dates and removing the time everywhere.
But this is actually unsafe too. Therefore, the current solution is to
not touch the time at all. As everything happens in the time zone of the
user, everything just works fine everywhere. Anyhow, this is not a good
and stable solution for the future and must be improved.

Due to the above descriptions, it was a bit unfortunate to fix the
original calculation code too. Though, it was attempted to keep the
changes minimal where possible.
The refactoring properties test suite got also more complicated due the
representation of "no bonus".

There are a few clean-up changes in this commit that remove utility
functions and method that have become obsolete and unused/dead-code.

Plenty of the related code is still absolutely ugly and horrible to
read. In the end, the commit should remain somewhat focused and limited
to fixing the bug. Plus some boyscouting. More to come.