feat(calculation): simplify data structure for children
First of all, there is no more meaning/purpose/reason to have the
possibility of an undefined birthdate for siblings. So it got removed
and allowed for a few simplifications. But more importantly it is more
precise and strict.

Furthermore, the `istBehindert` property was "improved" as flag, which
allows it to be missing (read: is optional). This improves the developer
experience and aligns with all the other flag properties.

Finally, some leftovers of the former `nummer` property were cleaned up.