fix(calculation): type safe access to planning data
Unfortunately, TypeScript does not properly communicate that indexing an
array of unknown length or unknown index can return an undefined value.
For the moment, this has been made more explicit by making the getter
signature aware of it, forcing users to follow.

In result a few mistakes were made. Some of them can't be properly fixed
without much effort. But the affected output data points were actually
obsolete and of no value. Therefore, this could be mostly fixed by
discarding this output. The remaining gap had to be closed by some
transformations in the refactoring property tests.

Because of this change, it is now more obvious that a planning might not
stretch for all possible months. That is fine and was actually already
used this way. In result, it is no more necessary to produce a full and
empty planning list on construction.