refactor: use more strict TypeScript for indexing
Unfortunately this is no default when setting TypeScript to be strict.
With this compiler option set, all objects and arrays are checked
for possible undefined values. It has been observed that it was/is
a common issue to have unsafe array indexing. Unfortunately, the type
systems was not helpful. This has changed now. Though, it is sometimes
not always easy to overcome the "c'mon this is obvious" and actually
convince the compiler this is correct. And don't forget, just because an
array has a certain length, this does not mean it actually has an entry
for a certain index within that range (JavaScript...). After all, this
will produce much more safe code.

Plenty of cases were simply fixed by adding optional handling. This
covers some pure type design in many cases. But it works solid for now
and often just represents the implicit handling before.
Some test utilities simply don't allow for a proper handling. So it is
usually just "enforced", sometimes with an additional expect statement.
Other cases have received some proper improvements to work with these
constraints to produce actual safe code.