refactor(calculation): remove getter for planning data
Since the last changes, this getter is solely logic is to shift the
indexes from one-based to zero-based. But this is actually a problem on
the caller side. If the caller wants to work with indices starting by
one, it should not put this responsibility on this class. It is rather
a hint to a flaw in the algorithm and data design.
This helps to simplify the class of the planning data.