refactor(calculation): release methods of financial data class
The goal is to simplify the input data structures of the algorithm. The
methods of this class are not necessary. One method is just a predicate
behind a getter. Another method was used at one single location only.
The last was a much more frequently used.
Anyhow, none of the methods were actually strongly bound to the class
itself. Actually they only read one single property values. They are
just associated functions that are needed for the algorithm. But there
are no good arguments to have them as class methods.
Due to the lack of a better idea how to locate this, one method has been
put along in the same module. It gets used by multiple other modules and
there is currently no well module structure for better placement.